updated to create schedules when passed
