+++ Foqos/Utils/AppBlockerUtil.swift
@@ -1,11 +1,6 @@
-import DeviceActivity
 import FamilyControls
 import ManagedSettings
 import SwiftUI
-
-extension DeviceActivityName {
-  static let daily = Self("daily")
-}
 
 class AppBlockerUtil {
   let store = ManagedSettingsStore(
@@ -44,29 +39,6 @@ class AppBlockerUtil {
     }
 
     store.application.denyAppRemoval = strict
-  }
-
-  func scheduleRestrictions(for profile: BlockedProfiles) {
-    // Configure a daily schedule from 7:30 PM to 8:00 PM
-    let intervalStart = DateComponents(hour: 19, minute: 30)
-    let intervalEnd = DateComponents(hour: 20, minute: 0)
-    let schedule = DeviceActivitySchedule(
-      intervalStart: intervalStart,
-      intervalEnd: intervalEnd,
-      repeats: true,
-    )
-
-    let center = DeviceActivityCenter()
-
-    // Stop any existing monitoring for the same name before starting again
-    center.stopMonitoring([.daily])
-
-    do {
-      try center.startMonitoring(.daily, during: schedule)
-      print("Scheduled restrictions from 7:30 PM to 8:00 PM daily")
-    } catch {
-      print("Failed to start monitoring: \(error.localizedDescription)")
-    }
   }
 
   func deactivateRestrictions() {
